+++ role.longHarvester.js
@@ -25,25 +25,24 @@ module.exports = {
         // const containers = creep.pos.findInRange(FIND_STRUCTURES, 10, {filter: {structureType: STRUCTURE_CONTAINER}});
         // console.log("containers:", containers);
 
-        // If creep already on duty, stay on current duty
-        if (creep.memory.onduty) {
-          upgrader.run(creep);
-          return;
-        }
-
         // If creep not on duty, find something to do
         let structure = creep.pos.findClosestByPath(FIND_MY_STRUCTURES, {
             // Filter results by appropriate structure type, as follow
             filter: structure => ((
+              // structure = creep.room.storage
               structure.structureType == STRUCTURE_STORAGE
               || structure.structureType == STRUCTURE_SPAWN
               || structure.structureType == STRUCTURE_EXTENSION
               || structure.structureType == STRUCTURE_TOWER)
-              && structure.energy < structure.energyCapacity) // structure low in energy
-            });
+              && structure.energy < structure.energyCapacity
+            ) // structure low in energy
+          });
 
-        // Backup to storage if all other structure are fully charged
-        if (!structure) structure = creep.room.storage;
+        // If no structure found, try find a storage unit
+        if (!structure) {
+          structure = creep.room.storage;
+          creep.say("Storage!");
+        }
 
         // If found a valid structure, transfer energy to it
         if (structure) {
@@ -51,15 +50,13 @@ module.exports = {
           if (creep.transfer(structure, RESOURCE_ENERGY) == ERR_NOT_IN_RANGE) {
             // if not in range, move closer
             creep.moveTo(structure, {
-            visualizePathStyle: { stroke: '#ffffff' }
+            visualizePathStyle: { stroke: '#7fffd4' } // aquamarine
             });
           }
         }
 
-        // If not target available, upgrade controller
+        // If not target available, try to upgrade controller
         if (!structure) {
-          // Keep creep on duty
-          creep.memory.onduty = true;
           upgrader.run(creep);
         }
 
@@ -80,20 +77,20 @@ module.exports = {
 
       if (creep.room.name == creep.memory.target) {
         // const source = creep.pos.findClosestByPath(FIND_SOURCES_ACTIVE);
-        const source = creep.room.find(FIND_SOURCES)[0];
+        const source = creep.room.find(FIND_SOURCES)[1];
 
         // Try harvesting from source
         if (creep.harvest(source) == ERR_NOT_IN_RANGE) {
           // Otherwise get closer to source
           creep.moveTo(source, {
-            visualizePathStyle: { stroke: '#ffffff' }
+            visualizePathStyle: { stroke: '#7fffd4' } // aquamarine
           });
         }
 
       // If not in target room, find exit and move towards it
       } else {
         const exit = creep.room.findExitTo(creep.memory.target);
-        creep.moveTo(creep.pos.findClosestByRange(exit));
+        creep.moveTo(creep.pos.findClosestByPath(exit));
       }
     }
   }