+++ role.upgrader.js
@@ -22,7 +22,7 @@ module.exports = {
 
     // Go recharging
     } else {
-      creep.recharge(true, true);
+      creep.recharge(true, false);
     }
   }
 }